refactor: add fileops, use fileops functions
